【录制框架】告警修改&修复xts用例失败

+++ interfaces/inner_api/native/av_common.h
@@ -172,7 +172,7 @@ struct WatermarkConfiguration {
     int32_t height = -1;
 };
 
-constexpr uint64_t MAX_WATERMARK_SIZE = 4 * 4096 * 4096;
+constexpr int32_t MAX_WATERMARK_SIZE = 4 * 4096 * 4096;
 constexpr int32_t WATERMARK_WIDTH_HEIGHT_MAX = 4096;
 constexpr int32_t WATERMARK_COUNT_MAX = 5;
 } // namespace Media
